requirements.txt
requests~=2.28.2
setuptools~=65.5.1
gitpython